Refactor part 1 to take battery size arg
For part 2 we need to change the number of cells to turn on so we need
to make this method more generic. This now loops the # of cell times to
select each and makes sure to update the reference for where to start
the next iteration as 1 to the right of the max selected value.

We need to be careful for each loop to only consider cells up to
(battery_size - index of cell in result) to ensure we can select the
remaining number of cells.