Solve day 5 part 2
For this part the problem asks for the count of distinct fresh IDs. To
solve this, we need to deduplicate (remove overlaps) from the input
ranges.

We do this by iterating over the ordered ranges provided and expanding
the "current" range as long as there is overlap. When there is no longer
overlap, we start a new range and continue expanding that one.