Solve Day 6 Part 1
We essentially get the input for this part as a transposed matrix once
we split the lines on whitespace. To solve the problem we transpose the
input matrix so that each line is an operation. We then just join the
lines using their last entry (the operator) and eval the result
expression.