Solve day 6 part 2
This is pretty gross, but I'm not going to optimize it any more.

For part 1 we could just split the lines to get the values and transpose
those numbers to get the expressions to evaluate. For part 2, we need to
do something similar except effectively "rotate" the entire input 90
degrees counter clockwise retaining column positions of characters and
then read line by line. This way we know blank lines separate each set
of operations.